refactors collections into folders
