+++ src/common/components/Dashboard/Projects.tsx
@@ -1,45 +1,43 @@
-import { GoFileMedia, GoDesktopDownload } from "react-icons/go";
-import Folder from './Folder';
-import type { CollectionType, ProjectType } from '@/common/types';
-import { clickHandlerTest } from '@/common/modules/utils';
+// import { GoFileMedia, GoDesktopDownload } from "react-icons/go";
+// import Folder from './Folder';
+// import type { CollectionType, ProjectType } from '@/common/types';
+// import { clickHandlerTest } from '@/common/modules/utils';
 import styles from '@/styles/components/Projects.module.scss';
 
-type Props = {
-  currentCollection: CollectionType,
-  projects: ProjectType[],
-  selectProject: React.MouseEventHandler,
-  createProject: React.MouseEventHandler,
-  downloadProjects: React.MouseEventHandler,
-};
+// type Props = {
+//   currentCollection: CollectionType,
+//   projects: ProjectType[],
+//   selectProject: React.MouseEventHandler,
+//   createProject: React.MouseEventHandler,
+//   downloadProjects: React.MouseEventHandler,
+// };
+// {
+//   // currentCollection,
+//   // projects,
+//   // selectProject,
+//   // createProject,
+//   // downloadProjects
+// }: Props
 
-export default function Projects({
-  currentCollection,
-  projects,
-  selectProject,
-  createProject,
-  downloadProjects
-}: Props) {
+export default function Projects() {
 
   return (
     <div className={styles.projects}>
+      <h3>PROJECTS</h3>
 
       {/* Create/Download Projects */}
-      <div className={styles.projectControls}>
-
+      {/* <div className={styles.projectControls}>
         <h2>{currentCollection.name ? currentCollection.name : 'ALL'}</h2>
-
-        {/* <div className={styles.buttons}> */}
         <button name='create-project' onClick={createProject}>
           <GoFileMedia />
         </button>
         <button name='download-projects' onClick={downloadProjects}>
           <GoDesktopDownload />
         </button>
-        {/* </div> */}
-      </div>
+      </div> */}
 
       {/* Project Folder List */}
-      <div className={styles.projectsList}>
+      {/* <div className={styles.projectsList}>
         {
           projects.length > 0 ?
             projects.map((project, i) => {
@@ -54,7 +52,7 @@ export default function Projects({
             })
             : null
         }
-      </div>
+      </div> */}
 
     </div>
   );
