+++ src/common/components/Dashboard/index.ts
@@ -1,0 +1,4 @@
+import Dashboard from './Dashboard';
+import Projects from './Projects';
+
+export { Dashboard, Projects };
