+++ src/common/components/ProjectForm/index.ts
@@ -1,0 +1,3 @@
+import ProjectForm from './ProjectForm';
+
+export { ProjectForm };
