+++ src/common/defaults/defaultFolder.ts
@@ -1,0 +1,8 @@
+import type { FolderType } from '@/common/types';
+
+const defaultFolder: FolderType = {
+  _id: '',
+  name: '',
+};
+
+export default defaultFolder;
