+++ src/common/defaults/index.ts
@@ -1,0 +1,7 @@
+import defaultFolder from './defaultFolder';
+import defaultProject from './defaultProject';
+
+export {
+  defaultFolder,
+  defaultProject,
+};
