--- src/common/types/CollectionType.ts
@@ -1,9 +1,0 @@
-// import { Types } from 'mongoose';
-
-type CollectionType = {
-  _id: string,
-  name: string,
-  projects: string[],
-};
-
-export default CollectionType;
