+++ src/common/types/FolderType.ts
@@ -1,0 +1,6 @@
+type FolderType = {
+  _id: string,
+  name: string,
+};
+
+export default FolderType;
