+++ src/styles/components/Dashboard.module.scss
@@ -1,0 +1,58 @@
+.dashboard {
+  border: $border-test-orange;
+  padding: 0.5rem;
+  margin: 0.25rem;
+  width: 25%;
+  display: flex;
+  flex-direction: column;
+
+  h1 {
+    margin: 0;
+  }
+
+  input, select {
+    width: 80%;
+  }
+
+  button {
+    border: none;
+    margin-left: 0.5rem;
+    margin-right: 0.5rem;
+    display: flex;
+    font-size: 2rem;
+  }
+
+  .dashboardMenu {
+    border: $border-test-blue;
+    padding: 0.5rem;
+  }
+
+  .newFolderForm {
+    border: $border-test-green;
+    font-size: 1rem;
+    display: flex;
+    align-items: center;
+
+    label {
+      display: none;
+    }
+  }
+
+  .selectFolderForm {
+    border: $border-test-green;
+    font-size: 1rem;
+    display: flex;
+    align-items: center;
+    
+    label {
+      display: none;
+    }
+  }
+
+  .createProjectForm {
+    border: $border-test-green;
+    font-size: 1rem;
+    display: flex;
+    align-items: center;
+  }
+}
