+++ src/styles/components/Folders.module.scss
