+++ src/styles/components/ProjectForm.module.scss
@@ -1,10 +1,13 @@
-#form {
-  // border: $border-test-orange;
-  // padding: $padding-stock;
-  height: 100%;
+.projectForm {
+  border: $border-test-orange;
+  margin: 0.25rem;
+  width: 75%;
   display: flex;
   flex-direction: column;
-  gap: 0.25rem;
+
+  h1 {
+    margin: 0;
+  }
 
   label {
     display: flex;
