+++ src/styles/globals.scss
@@ -13,7 +13,13 @@ html {
 body {
   margin: 0;
   font-family: $page-font;
-  font-size: 62.5%;
+  font-size: 100%;
+  height: 100vh;
+}
+
+#__next {
+  border: $border-stock;
+  height: 100%;
 }
 
 button {
@@ -26,7 +32,6 @@ button {
 }
 
 button:hover,
-.onclick:hover,
 input:hover,
 textarea:hover,
 select:hover {
@@ -34,37 +39,8 @@ select:hover {
   color: $active-color;
 }
 
-button.active {
-  color: $active-color;
-}
-
 main {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.project-div {
   border: $border-stock;
-  padding: $padding-stock;
   height: 100%;
   display: flex;
-  gap: 1rem;
-}
-
-.project-dashboard {
-  border: $border-stock;
-  padding: $padding-stock;
-  min-width: 20rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.project-form {
-  border: $border-stock;
-  padding: $padding-stock;
-  width: 100%;
-  gap: 1rem;
 }
